Started preparing for the actual testclass
