+++ Eduria/Eduria/Models/TestModel.cs
@@ -1,4 +1,4 @@
-﻿using Eduria.Models.Interfaces;
+﻿using Eduria.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,8 +8,22 @@ namespace Eduria.Models
 {
     public class TestModel
     {
-        public List<QAComboModel> QaComboModels;
+        public List<CombinedQuestionAnswer> CombinedQuestionAnswers;
 
 
+        public TestModel()
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool CheckAnswer(int id, int givenAnswer = 0, string givenAnswerString = "")
+        {
+            return true;
+        }
+
+        public void InsertQuestion(int id)
+        {
+            CombinedQuestionAnswers.Insert(1, new CombinedQuestionAnswer());
+        }
     }
 }
