+++ Eduria/Eduria/Services/QuestionAnswerService.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Eduria.Interfaces;
+
+namespace Eduria.Services
+{
+    public class QuestionAnswerService
+    {
+    }
+}
